Restructuring code structure
